+++ tests/hass-addon-sunsynk/test_run.py
@@ -1,5 +1,7 @@
 """Test run."""
 import logging
+import re
+from pathlib import Path
 from types import ModuleType
 from unittest.mock import patch
 
@@ -22,11 +24,45 @@ def run() -> ModuleType:
 def test_run(run):
     """Test Run."""
     assert not run.SENSORS
-    assert not run.OPTIONS.mqtt_host
+    assert not run.OPT.mqtt_host
 
     testargs = ["run.py", "host1", "passw"]
     with patch.object(run.sys, "argv", testargs):
         run.startup()
     assert run.SENSORS
-    assert run.OPTIONS.mqtt_host == "host1"
-    assert run.OPTIONS.mqtt_password == "passw"
+    assert run.OPT.mqtt_host == "host1"
+    assert run.OPT.mqtt_password == "passw"
+
+
+@pytest.mark.addon
+def test_versions(run):
+    """Test versions.
+
+    config.json - contains the HASS addon version
+    Dockerfile  - installs the specific sunsynk library from pypi
+    setup.py    - sunsynk library on pypi
+    """
+
+    def _get_version(filename, regex):
+        txt = Path(filename).read_text()
+        res = re.compile(regex).search(txt)
+        assert res, "version not found in setup.py"
+        return res.group(1)
+
+    v_setup = _get_version(
+        filename="setup.py",
+        regex=r'VERSION = "(.+)"',
+    )
+
+    v_docker = _get_version(
+        filename="hass-addon-sunsynk/Dockerfile",
+        regex=r"sunsynk==(.+)",
+    )
+
+    v_config = _get_version(
+        filename="hass-addon-sunsynk/config.json",
+        regex=r'"version": ".+-(.+)"',
+    )
+
+    assert v_setup == v_docker
+    assert v_setup == v_config