+++ hass-addon-sunsynk/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## **2021.9.15-0.0.4** - 2021-09-22
+
+- Enforce Sunsynk Serial number as SUNSYNK_ID
+- Library cleanup
+
 ## **2021.9.8** - 2021-09-21
 
 - MQTT: Clean retain config for unused sensors